rename back to debug.cpp and debug.h
